requirements.txt
django==3.0.3
djangorestframework==3.11.0
django-cors-headers==3.4.0
python-decouple==3.3
dj-database-url==0.5.0
gunicorn==20.0.4
psycopg2-binary>=2.8